+++ src/nota.h
@@ -5,19 +5,17 @@ class Nota {
 private:
     std::string numeStudent;  
     std::string numeMaterie;  
-    int nota;                  
+    double nota;                  
 
 public:
     Nota();
-    Nota(const std::string& numeStudent,
-         const std::string& numeMaterie,
-         int nota);
+    Nota(const std::string& numeStudent, const std::string& numeMaterie, double nota);
 
     std::string getNumeStudent() const;
     std::string getNumeMaterie() const;
-    int getNota() const;
+    double getNota() const;
 
     void setNumeStudent(const std::string& numeNou);
     void setNumeMaterie(const std::string& numeNou);
-    void setNota(int notaNoua);
+    void setNota(double notaNoua);
 };
